+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ## Added
 
 - Option to ignore autoaccept of Warband completed quests
-- DungeonTeleport now also adds a random available hearthstone (one pick per session) if you enable "Show Teleport Items and Toys"
+- DungeonTeleport now also adds a random available hearthstone - if you enable "Show Teleport Items and Toys"